requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.4
Pillow==6.2.1
pytz==2019.3
six==1.13.0
torch==1.3.1
torchvision==0.4.2
Werkzeug==0.16.0
